+++ packages/apptive_grid_core/CHANGELOG.md
@@ -3,6 +3,9 @@
 * Fixed parsing for `key`s in GridField
 * Fixed unauthorized retry for performing an ApptiveLink
 * Added custom header to calls where it was missing
+* Added support for all `ApptiveGridLayout` layouts
+### Breaking Change
+* `layout` in `loadEntities` is now `ApptiveGridLayout` instead of `String?`
 
 ## 0.10.0-alpha.4
 * Adjust GridField to include `key` and `schema`